pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flightfinder"
version = "0.1.0"
description = "Find cheapest A->X->B flights with a flexible city-break layover"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Flight-finder" }]
dependencies = [
    "fast-flights>=2.0",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "jinja2>=3.1",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
flightfinder = "flightfinder.cli:main"

[tool.setuptools.packages.find]
include = ["flightfinder*"]

[tool.setuptools.package-data]
flightfinder = ["templates/*.html", "static/*.css"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
